+++ Changelog.md
@@ -8,6 +8,7 @@
 - Fix: The death message of a player killed by a taser does not contain the player firing the taser
 - Fix: Crashes when rendering some modded block entities in a Projector, or using them as a disguise
 - Fix: The Sentry-/Mine Remote Access Tool screens do not show the item's custom name
+- Fix: Reinforced Doors can be closed by giving them a block update
 
 --------------------------Changelog for v1.9.6.1 of SecurityCraft--------------------------